Directory for Postgres stuff
About to get started refactoring adding a few basic things here that I've already written here. But as soon as I get settled in at the ATL, I'm going town on this directory like a maniac on speed — which I may or may not be — unleashing a monsoon of SQL files for almost everything necessary for the FSTDT database. (If this stuff is conventionally kept somewhere else, plz move it as appropriate, or better yet, yell at me to do it. Also feedback if this description is bad, etc.)